CFY-6133. Update log/event schema (#435)
Additional data that wasn't stored in the logs/events table before this PR is now part of the logs/events tables schema like event type or logger. JSON data was initially considered to use the same schema as in elasticsearch, but it was finally discarded to make things simple and also because it wasn't really needed.

General refactoring of the code that takes care of the `logstash-output-jdbc` plugin and the postgresql jdbc driver installation.