CFY-6091. Create log tables in rest service (#441)
* CFY-6091 Remove logs/events table creation code

From now on, the database tables will be created when the restservice is
configured

* CFY-6091 Update insert queries to include execution_fk

Logs/events derive from executions, so they need to insert a foreign key
to them.

* CFY-6091 Remove message_vector from insert queries

The TSVECTOR data type is PosgreSQL specific, so it will be removed
for now to keep the ability to switch between different database
engines. An alternative approach to full-text search will be evaluated
later.