CFY-6090 Insert null values instead of empty strings
The logging code checks if node_id/operation is None to decide whether
to include it or not. NULLIF is helpful here to avoid inserting the
empty string.

Another solution could have been using ruby code to set values to nil by
default. However, the logstash-output-jdbc plugin doesn't work well with
that because when a value is NULL it inserts the substitution string
instead.