CFY 6090. Add missing columns to events (#474)
* CFY-6090 Insert the operation field into events

* CFY-6090 Set empty string as the default value for operation

Without this default value, the query that inserts the operation value
in the events table might fail because not all events have this field
set.

* CFY-6090 Insert the node_id field into events

* CFY-6090 Set empty string as the default value for node_id

Without this default value, the query that inserts the node_id value
in the events table might fail because not all events have this field
set.

* CFY-6090 Insert null values instead of empty strings

The logging code checks if node_id/operation is None to decide whether
to include it or not. NULLIF is helpful here to avoid inserting the
empty string.

Another solution could have been using ruby code to set values to nil by
default. However, the logstash-output-jdbc plugin doesn't work well with
that because when a value is NULL it inserts the substitution string
instead.

* CFY-6090 Updated circle CI environment to m13

Updated depencencies that are pip installed in circle CI to m13

* CFY-6090 Updated travis CI environment to m13

Updated dependencies that are pip installed in travis CI to m13