Add the / vhost to rabbitmq definitions.json
All our exchanges require that vhost; make sure it is created
always when the exchanges are created, by declaring it in the
definitions file that declares the exchanges.

In normal operation, the / vhost is created when installing
rabbitmq, but it can sometimes be deleted (eg. when resetting
rabbitmq's storage); after that, the exchanges were created as in
the definitions file, but the vhost had to be created by hand.