Updated to more correct relationship
At runtime, it doesn't matter whether the port is `connected_to` the network or `contained_in`, because the code looks for any kind of a relationship. So the original code works. However, it serves as a bad example because this relationship between a port and a network is *incorrect*. Among other things, a `contained_in` relationship between a port and a network would prevent the user from using scaling groups to include a Server and a Port in the same scaling group; that wouldn't pass validation, because `Port` would be `contained_in` a network and thus `Server` and `Port` have an ancestry conflict.