CFY-6576 Make IPSetter and stage depend on manager_configuration
Every node that uses fabric (for .py scripts) needs to have a
depends_on relationship to the manager_configuration, because that's
where the host_string is taken from, for the fabric env.

Without this change, some nodes can be created before
manager_configuration, which will lead to an empty host_string in
their fabric inputs, and an error message from fabric.