CFY 6479. Serialize error causes (#540)
* CFY-6479 Install json_encode filter plugin

This is needed to serialize the `error_causes` json object to a string
before running the insert query. Otherwise, what gets store into the
database is a ruby style serialization of the object.

* CFY-6479 Serialize `task_error_causes` to a string

The database will store this as a string for now to avoid
incompatibilities between PostgreSQL and SQLite in the testing
environment.