CFY-6908 Add the "restore stage snapshot" script (#597)
* Add env= and stdin= to run()/sudo() helpers

Supporting those arguments can simplify the parts of code that use
a shell to pass stdin or env vars

* Add to sudoers: also allow sudoing to users other than root

When we want to sudo to some other user than root, eg. the stage user,
this will allow us to specify exactly what sudo commands are allowed

* add_entry_to_sudoers: use Popen/run()/sudo() commands instead of bash

Using a shell isn't necessary anymore, now that we can pass stdin and
env directly

* Add the "restore stage snapshot" script

Mgmtworker can't restore stage files directly, because those are
only writable by the stage user. Instead, we'll add a script that
replaces the snapshot files, and allow mgmtworker to sudo that
script as the stage user