CFY-7215 Use Type=oneshot for IPSetter
For type=oneshot services, the services depended on it, will wait for
the oneshot service to finish before they run.

Without this, IPSetter is the default "simple" type, which means that
nginx might start running BEFORE ipsetter finishes, and then nginx
will be using the old certificate, before it was replaced.