Split generate_ca_cert and generate_internal_ssl_cert
We'll be working in terms of those 2 functions: generate_ca_cert
and generate_internal_ssl_cert.

Only the CA cert needs to also be in the .p12 format.

Only the internal cert needs the subjectAltNames (ip/dn) in it. That
call is not required for the CA cert.

Also, the signature of generate_internal_ssl_cert is changed to
take in the subject commonName ("CN") separately, instead of just
assuming the only IP passed should be it.