Create the certs in configure_manager
The certs are used for several things (nginx, rabbitmq, and sending
them to the agents). It doesn't quite make sense to create them in
nginx's scripts. Instead, let's create the certs in the
manager_configuration node, which is really the earliest they can be
created at.

If, in the future, we decide to use one cert for nginx, and another
for rabbitmq, each service can create its own cert, but the CA would
still be manager-wide, so should then still be created here, on the
manager level.