Run set_manager_ips.py in preconfigure, not postconfigure
To do the previous commit (create certs in configure_manager), the
internal_rest_host runtime prop needs to be available there.
That property is set by the set_manager_ips script, which now needs to
run BEFORE configure_manager, so we can set it to run in preconfigure.

Turns out there's no real reason for set_manager_ips to run in
postconfigure. The only thing that actually has to be changed is that
the default internal port needs to be set in it as well. This change
makes sense IMO, because now we set the port in the same place we
set the host.

To do that, we move the constant from utils.py to the script. It's
still only used in one place, to be put into the runtime props.