Create the cloudify user in configure_manager
While we're at it, there's no reason for the user (cfyuser) to be
created in the ipsetter scripts. They are unrelated to that.

Instead, let's create that user at the manager configuration stage.

This is required here because we also create the certs there, and
will chown the related files (metadata) to cfyuser, so the user must
exist by that time.