IPSetter can only recreate the cert
IPSetter's create-internal-ssl-certs.py used to always take the
new address as an argument, and recreate the cert using that.

Instead, let's allow it to recreate the cert with the previous
internal_rest_host, so that certs can be re-created without changing
the ips in them.

This means just running the script will create a new cert signed by
the current CA, using the same addresses. And passing an argument
will create a new cert signed by the current CA, using the passed in
address (as before).