CFY-7190 Use a separate CA cert (#652)
* Cert path constants

We're now working with 3 main sets of paths:
 - the internal cert: /etc/cloudify/ssl/cloudify_internal_{cert,key}.pem
 - the ca cert: /etc/cloudify/ssl/cloudify_internal_ca_{cert,key}.pem
 - and the .p12 version of the CA for riemann

* Split generate_ca_cert and generate_internal_ssl_cert

We'll be working in terms of those 2 functions: generate_ca_cert
and generate_internal_ssl_cert.

Only the CA cert needs to also be in the .p12 format.

Only the internal cert needs the subjectAltNames (ip/dn) in it. That
call is not required for the CA cert.

Also, the signature of generate_internal_ssl_cert is changed to
take in the subject commonName ("CN") separately, instead of just
assuming the only IP passed should be it.

* Create the certs in configure_manager

The certs are used for several things (nginx, rabbitmq, and sending
them to the agents). It doesn't quite make sense to create them in
nginx's scripts. Instead, let's create the certs in the
manager_configuration node, which is really the earliest they can be
created at.

If, in the future, we decide to use one cert for nginx, and another
for rabbitmq, each service can create its own cert, but the CA would
still be manager-wide, so should then still be created here, on the
manager level.

* Run set_manager_ips.py in preconfigure, not postconfigure

To do the previous commit (create certs in configure_manager), the
internal_rest_host runtime prop needs to be available there.
That property is set by the set_manager_ips script, which now needs to
run BEFORE configure_manager, so we can set it to run in preconfigure.

Turns out there's no real reason for set_manager_ips to run in
postconfigure. The only thing that actually has to be changed is that
the default internal port needs to be set in it as well. This change
makes sense IMO, because now we set the port in the same place we
set the host.

To do that, we move the constant from utils.py to the script. It's
still only used in one place, to be put into the runtime props.

* Create the cloudify user in configure_manager

While we're at it, there's no reason for the user (cfyuser) to be
created in the ipsetter scripts. They are unrelated to that.

Instead, let's create that user at the manager configuration stage.

This is required here because we also create the certs there, and
will chown the related files (metadata) to cfyuser, so the user must
exist by that time.

* Use the metadata file to store networks IPs

To re-generate the certs later, we need to store the ips/names that
need to go into subjectAltNames.

For that, we'll use the metadata file. We'll store the networks in
there, and also the primary internal_rest_host (which will be
the one changed by ipsetter).

* Replace the internal cert with the CA cert in services that use it

Services that only use the cert to verify a server should use the
CA cert.

This is mostly just a formality

* Add params to the docstring in _generate_ssl_certificate

* _generate_ssl_cert: move the configfile parts to a separate function

* Whoops. Create the user BEFORE creating the certs.

Fixing a rebase fail.