CFY-7220 A way to recreate the certs with the current CA (#653)
* Always install the IPSetter, but not always enable it

We will be using the script from IPSetter to re-create the internal
cert in a HA environment. Let's always put the scripts there, so
that we can use them in a cluster, but only enable them if the
flag is set.

* IPSetter can only recreate the cert

IPSetter's create-internal-ssl-certs.py used to always take the
new address as an argument, and recreate the cert using that.

Instead, let's allow it to recreate the cert with the previous
internal_rest_host, so that certs can be re-created without changing
the ips in them.

This means just running the script will create a new cert signed by
the current CA, using the same addresses. And passing an argument
will create a new cert signed by the current CA, using the passed in
address (as before).