Replace deploy_or_generate_external_ssl_cert with deploy_ssl_cert
We are going to use a "deploy" function for internal certs as well,
so it makes sense to have this as a more fine-grained building block.

And anyway, a "do_x_or_y" function is somewhat ugly, especially if it's
only used once. We can just move the check out to the call site.