Deploy the external cert using deploy_blueprint_resource calls
The deploy_ssl_cert function had very little of its own functionality,
so we can drop that layer and use deploy_blueprint_resource directly.

As an added bonus, now the case of only one of (cert, key) being
provided is handled better (an error is thrown with a descriptive
message, instead of failing someplace further down the road).