CFY-7236 Allow passing internal certs in inputs (#693)
* Fix flake

Those changes are unrelated, but i love having the file be clean on
a strict linter setting, They're very trivial.

* Replace deploy_or_generate_external_ssl_cert with deploy_ssl_cert

We are going to use a "deploy" function for internal certs as well,
so it makes sense to have this as a more fine-grained building block.

And anyway, a "do_x_or_y" function is somewhat ugly, especially if it's
only used once. We can just move the check out to the call site.

* Deploy the external cert using deploy_blueprint_resource calls

The deploy_ssl_cert function had very little of its own functionality,
so we can drop that layer and use deploy_blueprint_resource directly.

As an added bonus, now the case of only one of (cert, key) being
provided is handled better (an error is thrown with a descriptive
message, instead of failing someplace further down the road).

* Instead of always generating the internal CA + cert, try copying it

This will allow providing those certs by the user.

* Allow passing the internal cert, and internal CA, in inputs

The necessary bookkeeping - add properties, inputs, etc

* Implement a private _deploy_cert_and_key function

Of course, we still need to have the try/except logic on deploy.
Let's implement that. And, since we always try to deploy both the key
and the cert, let's do both in the same call.

* Rename *_cert to *_certificate

For consistency with rest_certificate

* nginx preconfigure: move out target_runtime_props to module level

* Factor out a _generate_external_cert

* A few comments about internal + CA certs

* Add empty defaults for the new inputs

* It's the path that can be empty, not the input name